refactor(group local storage): polyfill clear on asyncstorageDown
